test(layerize): restructuring test folder
restructuring test folder

feat #20
