+++ lib/index.js
@@ -413,6 +413,12 @@ class Layerize {
             unlock = await setLock(`LAYERIZE:LOCK:INSTALL:${schemaName.toUpperCase()}`, 60000);
 
             /**
+             * delete all related schemas caches
+             */
+            let layers = this.layers({ schemaName });
+            await layers.clearAllTablesCache();
+
+            /**
              * drop database schema and tables
              */
             await this.db.schema.raw(`DROP SCHEMA IF EXISTS ${schemaName} CASCADE`);
@@ -421,10 +427,6 @@ class Layerize {
              * delete all related ES schemas
              */
             await this.es.indices.delete({ index: `${schemaName}~*` });
-
-            /**
-             * TODO: delete all related schemas caches
-             */
 
             this.dbSchemas = await this.loadSchemaListFromDB();
