+++ lib/services/layers.js
@@ -58,7 +58,7 @@ class Layers {
          * Name of the schema being used within the layer
          * @member {string}
          * */
-        this.schemaName = name || '';
+        this.schemaName = name;
 
         /**
          * The initialized knex instance to use
@@ -160,17 +160,7 @@ class Layers {
          * A instance of the 'database' member that has been initiated with the schemaName
          * @member {knex}
          * */
-        this.dbSchema = null;
-
-        if (this.schemaName !== '') {
-
-            this.dbSchema = this.database.withSchema(this.schemaName).clone();
-
-        } else {
-
-            this.dbSchema = this.database.withSchema('').clone();
-
-        }
+        this.dbSchema = this.database.withSchema(this.schemaName).clone();
 
     }
 
@@ -274,7 +264,9 @@ class Layers {
             }
 
             const cacheKey = this.cache.key(this.cacheKey, table, '*');
-            return await this.cache.deleteByPattern(cacheKey);
+            await this.cache.deleteByPattern(cacheKey);
+
+            return { success: true };
 
         } catch (error) {
 
@@ -417,9 +409,15 @@ class Layers {
 
                     if (statement.removeFromCache) {
 
+                        if (typeof statement.cache === 'undefined') {
+
+                            statement.cache = [];
+
+                        }
+
                         for (let x = 0; x < result.rows.length; x++) {
 
-                            let key = result.row[statement.primaryKey];
+                            let key = result.rows[x][statement.primaryKey];
                             statement.cache.push(this.cache.key(this.cacheKey, statement.table, key));
 
                         }
@@ -1635,7 +1633,7 @@ class Layers {
 
                     if (!Array.isArray(fields)) {
 
-                        fields = fields.split(',');
+                        fields = this.__toArray(fields, 4);
 
                     }
 
@@ -1726,7 +1724,6 @@ class Layers {
     /**
      * Updates a single full records into the data layer
      * @param {string} table - name of the table
-     * @param {string} key - primary key of the record
      * @param {object} data - the full data of the record being updated
      * @param {object=} options - available options
      * @param {string|array} [options.fields=*] - fields to be returned
@@ -1737,7 +1734,7 @@ class Layers {
      * @returns {Promise<object>} If layer IS NOT a transaction and returnRecord is 'true', then it will return the reocrd object
      * @returns {Promise<success>} If layer IS NOT a transaction and returnRecord is 'false', then it will return a success object
      */
-    async update (table = '', key = '', data = {}, { returnRecord = false, ignoreReadOnly = [], fields = '', includes = '' } = {}) {
+    async update (table = '', data = {}, { returnRecord = false, ignoreReadOnly = [], fields = '', includes = '' } = {}) {
 
         this.debug('update()');
         try {
@@ -1750,6 +1747,14 @@ class Layers {
 
                 let column = schema.timestamps[i];
                 data[column.name] = 'now()';
+
+            }
+
+            let key = data[schema.primaryKey];
+
+            if (typeof key === 'undefined') {
+
+                throw new Error('When using update(), the object must contain the primary key with all other properties.');
 
             }
 
@@ -1784,7 +1789,16 @@ class Layers {
             } else {
 
                 let results = await this.__nonTransactionESCacheProcessing({ type: 'update', table, keys: [ key ], schema, statement, returnRecords: returnRecord, fields, includes });
-                return results[0];
+
+                if (returnRecord) {
+
+                    return results[0];
+
+                } else {
+
+                    return results;
+
+                }
 
             }
 
@@ -2474,7 +2488,7 @@ class Layers {
                             }
 
                             objArgName = objArgName.replace(/\,/gi, '.');
-                            aryArgValue = objArgValue.split(' ');
+                            aryArgValue = this.__toArray(objArgValue, 3);
 
                             switch (objArg.sep) {
 
@@ -2762,7 +2776,7 @@ class Layers {
 
                 } else {
 
-                    includes = includes.split(',');
+                    includes = this.__toArray(includes, 4);
 
                 }
 
@@ -3529,17 +3543,10 @@ class Layers {
      * @param {number} type - (0) split by |<br> (1) split by :<br> (2) split by () && ||<br> (3) split by space<br> (4) split by comma
      * @returns {array} returns array from split string
      */
-    __toArray (str, type) { //string into array
+    __toArray (str, type = 4) { //string into array
 
-        if (!type) {
-
-            type = 0;
-
-        }
         switch (type) {
 
-            case 0: // |
-                return str.match(/(?:[^\|]|\|\|)+/g);
             case 1: // :
                 // return str.split(/(?:[^[:]]|:)+/g);
                 return str.split(/:/g);
@@ -3551,8 +3558,6 @@ class Layers {
             case 4: //comma
                 str = this.__removeAllSpaces(str);
                 return str.split(',');
-            default:
-                return [];
 
         }
