+++ package.json
@@ -5,7 +5,7 @@
   "main": "lib/index.js",
   "scripts": {
     "test": "mocha --exit",
-    "test-coverage": "nyc mocha --exit",
+    "test-coverage": "nyc npm test && nyc report --reporter=html",
     "test-travis": "nyc npm test && nyc report --reporter=text-lcov | coveralls",
     "test-debug": "DEBUG=layerize* mocha",
     "commit": "git cz",
@@ -39,7 +39,7 @@
   "bugs": {
     "url": "https://github.com/LayerizeJS/layerize/issues"
   },
-  "homepage": "https://github.com/LayerizeJS/layerize#readme",
+  "homepage": "https://layerizejs.github.io/",
   "devDependencies": {
     "@enspire/eslint-config-enspire": "^1.1.1",
     "coveralls": "^3.0.2",
@@ -51,6 +51,7 @@
     "lint-staged": "^8.1.0",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
+    "mocha-prepare": "^0.1.0",
     "validate-commit-msg": "^2.14.0"
   },
   "dependencies": {
