+++ test/mocha.opts
@@ -1,0 +1,2 @@
+--recursive
+--require ./test/bootstrap
