+++ wahoo.md
@@ -1,0 +1,66 @@
+# WahooRMS TODO
+
+- [ ] Setup Github
+- [ ] Setup Jira
+- [ ] Setup Travis CI
+- [ ] Setup Coveralls CI (code coverage)
+- [ ] Setup Commitizens
+- [ ] Generate Release Notes / Change Logs
+- [ ] Enforce eslint before commit
+- [ ] LayerizeJS
+- [ ] Document and Diagram organization flows related to code development (features, issues, code)
+- [ ] i18n (i18next & locize.com)
+- [ ] WahooRMS services
+    - [ ] Settings
+        - [ ] Configuration
+        - [ ] Notifications
+        - [ ] Integration
+            - [ ] Queues
+            - [ ] Webhooks
+            - [ ] Scripts
+            - [ ] Jobs
+                - [ ] Events
+                - [ ] Scheduled
+        - [ ] Calendars
+    - [ ] Users
+    - [ ] Permissions
+    - [ ] Sites
+    - [ ] Items
+    - [ ] Inventory
+        - [ ] Purchasing
+        - [ ] Receiving
+        - [ ] Transfers
+        - [ ] Adjustments
+        - [ ] Cycle counts
+    - [ ] Customers
+    - [ ] Orders
+    - [ ] Fulfillments
+        - [ ] Picking/Pack
+        - [ ] Shipping
+    - [ ] Reporting
+    - [ ] Services Testing
+    - [ ] Services Documentation
+    - [ ] Testing
+    - [ ] 100% Code Coverage
+- [ ] Socket.io Services
+- [ ] GraphQL
+- [ ] RESTful API
+- [ ] UI Backoffice (MarkoJS)
+    - [ ] Dashboard
+    - [ ] Orders
+    - [ ] Fulfillments
+    - [ ] Customers
+    - [ ] Inventory
+    - [ ] Items
+    - [ ] Reporting
+    - [ ] Settings
+    - [ ] UI Testing
+    - [ ] UI Documentation
+    - [ ] Testing
+    - [ ] 100% Code Coverage
+- [ ] Point of Sale (Flutter.io) 
+    - [ ] Testing
+    - [ ] 100% Code Coverage
+- [ ] Marketing Site
+    - [ ] Testing
+    - [ ] 100% Code Coverage
